==> Backend/app/preprocessing/1. remove_columns.ipynb <==
import pandas as pd

df = pd.read_csv(r'../data/github_dataset.csv')

df.head()

# deleting the unwanted columns using drop()
df.drop(['Username','Profile URL','Top Languages','Following', 'Total Forks', 'Public Gists','Bio','Public Email','Forked Repositories','Last Activity Date'], axis=1,  inplace=True)

df.head()

df.to_csv('../data/cleaned_dataset_step_1.csv', index=False)
print("Cleaned dataset saved as '../data/cleaned_dataset_step_1.csv'")
