==> Backend/app/preprocessing/3. handle_missing_values-2.ipynb <==
import pandas as pd

# Load the CSV with correct parsing options
df = pd.read_csv(r'../data/cleaned_dataset_step_2.csv')

df.head()
print(len(df))

missing_values = df.isnull().sum()
print("Missing values per column: ")
print(missing_values)

print(df.dtypes)

# Filter repositories where 'Most Used Language' is null and 'Total Repository Size (KB)' is 0
empty_repos = df[(df['Most Used Language'].isnull()) & 
                 (df['Total Repository Size (KB)'] == 0)]

# Count of those repos
print(f"Number of repositories with no most used language and size 0: {len(empty_repos)}")

# Drop rows where 'Most Used Language' is null and 'Total Repository Size (KB)' is 0
df = df[~((df['Most Used Language'].isnull()) & 
          (df['Total Repository Size (KB)'] == 0))]

# Verify if the rows are removed
print(len(df))


missing_values = df.isnull().sum()
print("Missing values per column after dropping Empty Repo with no Most Used Language: ")
print(missing_values)

unique_values = df['Most Used Language'].unique()

print(len(unique_values))

value_counts = df['Most Used Language'].value_counts()

print(value_counts)

# checking if the missing values in Repositories Count, Total Repository Size (KB), Total Stars and Followers are the same
missing_rows = df[df['Country'].isnull() & df['Most Used Language'].isnull()]
print("Total count of missing Top Repo, Repo count, total stars and followers: ",len(missing_rows))

df = df[~((df['Most Used Language'].isnull()) & 
          (df['Country'].isnull()))]

# Verify if the rows are removed
print(len(df))

df.to_csv('../data/cleaned_dataset_step_3.csv', index=False)
