==> Backend/app/preprocessing/5. handle_missing_values-3.ipynb <==
import pandas as pd

df = pd.read_csv(r'../data/manually_cleaned_dataset_step_5.csv')

df.tail()

missing_values = df.isnull().sum()
print("Missing values per column: ")
print(missing_values)

df['Country'] = df['Country'].fillna('Unknown')

df['Country'].unique()

missing_values = df.isnull().sum()
print("Missing values per column: ")
print(missing_values)

# Filter out rows where Country is "Unknown"
filtered_df = df[df['Country'] != 'Unknown']
print(len(filtered_df))

missing_values = df.isnull().sum()
print("Missing values per column: ")
print(missing_values)

df = df.dropna(subset=['Most Used Language'])

missing_values = df.isnull().sum()
print("Missing values per column: ")
print(missing_values)

df['Unique Topics'] = df['Unique Topics'].fillna('Unknown')

missing_values = df.isnull().sum()
print("Missing values per column: ")
print(missing_values)

df.to_csv('../data/final_cleaned_data.csv', index=False)
print("Results saved to '('../data/final_cleaned_data.csv'.")
