==> Backend/app/preprocessing/6. datatypes_handling.ipynb <==
import pandas as pd

df = pd.read_csv('../data/final_cleaned_data.csv')

missing_values = df.isnull().sum()
print("Missing values per column: ")
print(missing_values)

check = df['Unique Topics'] == 'Unknown'
x = df[check]
print(len(x))

print(df.dtypes)

# Convert Country and Most Used Language to string
df['Country'] = df['Country'].astype('string')
df['Most Used Language'] = df['Most Used Language'].astype('string')

# Convert 'Unique Topics' to a list of strings (assuming it's a string of topics separated by commas)
df['Unique Topics'] = df['Unique Topics'].apply(lambda x: x.split(',') if isinstance(x, str) else [])

# Check the updated data types
print("Updated DataFrame Data Types:")
print(df.dtypes)

check = df['Unique Topics'] == 'Unknown'
x = df[check]
print(len(x))

df.to_csv('../data/finalized_dataset.csv', index=False)
print("Results saved to '('../data/finalized_dataset.csv'.")
